further refactorings
