+++ grails-app/domain/de/laser/Links.groovy
@@ -1,13 +1,13 @@
 package de.laser
 
-
+import com.k_int.kbplus.GenericOIDService
 import de.laser.exceptions.CreationException
 import org.grails.orm.hibernate.cfg.GrailsHibernateUtil
 
+import javax.persistence.Transient
+
 class Links {
 
-    def contextService
-    def springSecurityService
     def genericOIDService
 
     Long id
@@ -51,19 +51,17 @@ class Links {
 
     /**
      * Constructor for a Subscription/License linking. Parameters are specified in a {@link Map}.
-     * @param configMap - contains the parameters. The source and destination are expected on source and destination; the determination of object type is done in the setter {@link #setSourceAndDestination() setSourceAndDestination()}
+     * @param configMap - contains the parameters. The source and destination are expected on source and destination; the determination of object type is done in the setter {@link #setSourceAndDestination setSourceAndDestination()}
      * @return the persisted linking object
      * @throws CreationException
      */
     static Links construct(Map<String, Object> configMap) throws CreationException {
-        withTransaction {
-            Links links = new Links(owner: configMap.owner, linkType: configMap.linkType)
-            links.setSourceAndDestination(configMap.source,configMap.destination)
-            if (links.save())
-                links
-            else {
-                throw new CreationException(links.errors)
-            }
+        Links links = new Links(owner: configMap.owner, linkType: configMap.linkType)
+        links.setSourceAndDestination(configMap.source,configMap.destination)
+        if (links.save())
+            links
+        else {
+            throw new CreationException(links.errors)
         }
     }
