+++ grails-app/services/de/laser/LinksGenerationService.groovy
@@ -7,13 +7,19 @@ import de.laser.exceptions.CreationException
 import de.laser.helper.RDConstants
 import de.laser.helper.RDStore
 import grails.gorm.transactions.Transactional
+import grails.web.servlet.mvc.GrailsParameterMap
+import org.springframework.context.MessageSource
 import org.springframework.context.i18n.LocaleContextHolder
 
 @Transactional
 class LinksGenerationService {
 
+    static final int STATUS_OK = 0
+    static final int STATUS_ERROR = 1
+
     GenericOIDService genericOIDService
-    def messageSource
+    MessageSource messageSource
+    ContextService contextService
 
     LinkedHashMap<String,List> generateNavigation(context) {
         List prevLink = []
@@ -144,9 +150,79 @@ class LinksGenerationService {
      *
      * @return false if manipulation was successful, a string describing the error otherwise
      */
-    def createOrUpdateLink(Map<String,Object> configMap) {
-        def errors = false
-
+    Map<String,Object> createOrUpdateLink(GrailsParameterMap params) {
+        Locale locale = LocaleContextHolder.getLocale()
+        Map<String,Object> result = [institution:contextService.org]
+        //error when no pair is given!
+        params.keySet().each {
+            if(it.contains("pair_")) {
+                def pairCheck = params.get(it)
+                if(!pairCheck) {
+                    result.error = messageSource.getMessage('default.linking.noLinkError',null,locale)
+                    [result:result,status:STATUS_ERROR]
+                }
+            }
+        }
+        //distinct between insert and update - if a link id exists, then proceed with edit, else create new instance
+        Map<String,Object> configMap = [owner:result.institution]
+        //perspectiveIndex 0: source -> dest, 1: dest -> source
+        if(params.link) {
+            configMap.link = genericOIDService.resolveOID(params.link)
+            if(params.commentID)
+                configMap.comment = genericOIDService.resolveOID(params.commentID)
+            if(params["linkType_${configMap.link.id}"]) {
+                String linkTypeString = params["linkType_${configMap.link.id}"].split("§")[0]
+                int perspectiveIndex = Integer.parseInt(params["linkType_${configMap.link.id}"].split("§")[1])
+                configMap.linkType = genericOIDService.resolveOID(linkTypeString)
+                configMap.commentContent = params["linkComment_${configMap.link.id}"].trim()
+                if(perspectiveIndex == 0) {
+                    configMap.source = genericOIDService.resolveOID(params.context)
+                    configMap.destination = genericOIDService.resolveOID(params["pair_${configMap.link.id}"])
+                }
+                else if(perspectiveIndex == 1) {
+                    configMap.source = genericOIDService.resolveOID(params["pair_${configMap.link.id}"])
+                    configMap.destination = genericOIDService.resolveOID(params.context)
+                }
+            }
+            else if(!params["linkType_${configMap.link.id}"]) {
+                result.error = messageSource.getMessage('default.linking.linkTypeError',null,locale)
+                [result:result,status:STATUS_ERROR]
+            }
+        }
+        else {
+            if(params["linkType_new"]) {
+                String linkTypeString = params["linkType_new"].split("§")[0]
+                int perspectiveIndex = Integer.parseInt(params["linkType_new"].split("§")[1])
+                configMap.linkType = genericOIDService.resolveOID(linkTypeString)
+                configMap.commentContent = params.linkComment_new
+                if(perspectiveIndex == 0) {
+                    configMap.source = genericOIDService.resolveOID(params.context)
+                    configMap.destination = genericOIDService.resolveOID(params.pair_new)
+                }
+                else if(perspectiveIndex == 1) {
+                    configMap.source = genericOIDService.resolveOID(params.pair_new)
+                    configMap.destination = genericOIDService.resolveOID(params.context)
+                }
+                def currentObject = genericOIDService.resolveOID(params.context)
+                List childInstances = currentObject.getClass().findAllByInstanceOf(currentObject)
+                if(childInstances) {
+                    configMap.contextInstances = childInstances
+                    def pairObject = genericOIDService.resolveOID(params.pair_new)
+                    configMap.pairInstances = pairObject.getClass().findAllByInstanceOf(pairObject)
+                }
+            }
+            else if(params["linkType_sl_new"]) {
+                String linkTypeString = params["linkType_sl_new"].split("§")[0]
+                configMap.linkType = genericOIDService.resolveOID(linkTypeString)
+                configMap.commentContent = params.linkComment_sl_new
+                configMap.source = genericOIDService.resolveOID(params.pair_sl_new)
+                configMap.destination = genericOIDService.resolveOID(params.context)
+            }
+            else if(!params["linkType_new"] && !params["linkType_sl_new"]) {
+                result.error = messageSource.getMessage('default.linking.linkTypeError',null,locale)
+                [result:result,status:STATUS_ERROR]
+            }
+        }
         Doc linkComment
         if(configMap.comment instanceof Doc)
             linkComment = (Doc) configMap.comment
@@ -183,12 +259,12 @@ class LinksGenerationService {
             }
             catch (CreationException e) {
                 log.error( e.toString() )
-                errors = messageSource.getMessage('default.linking.savingError',null, LocaleContextHolder.getLocale())
+                result.error = messageSource.getMessage('default.linking.savingError',null, locale)
+                [result:result,status:STATUS_ERROR]
             }
         }
-
         if(link) {
-            link.linkType = configMap.linkType
+            link.linkType = (RefdataValue) configMap.linkType
             link.setSourceAndDestination(configMap.source, configMap.destination)
             if(linkComment) {
                 if(configMap.commentContent.length() > 0) {
@@ -197,8 +273,8 @@ class LinksGenerationService {
                 }
                 else if(configMap.commentContent.length() == 0) {
                     DocContext commentContext = DocContext.findByOwner(linkComment)
-                    if(commentContext.delete(flush:true))
-                        linkComment.delete(flush:true)
+                    if(commentContext.delete())
+                        linkComment.delete()
                 }
             }
             else if(!linkComment && configMap.commentContent.length() > 0) {
@@ -210,15 +286,17 @@ class LinksGenerationService {
                 }
                 else {
                     log.error(linkComment.errors.toString())
-                    errors = messageSource.getMessage('default.linking.savingError',null, LocaleContextHolder.getLocale())
+                    result.error = messageSource.getMessage('default.linking.savingError',null, locale)
+                    [result:result,status:STATUS_ERROR]
                 }
             }
         }
         else if(link && link.errors) {
             log.error(link.errors.toString())
-            errors = messageSource.getMessage('default.linking.savingError',null, LocaleContextHolder.getLocale())
+            result.error = messageSource.getMessage('default.linking.savingError',null, locale)
+            [result:result,status:STATUS_ERROR]
         }
-        errors
+        [result:result,status:STATUS_OK]
     }
 
     boolean deleteLink(String oid) {
